+++ quantforge/__init__.py
@@ -750,6 +750,7 @@ from .weighted_stats import (
 from .fenwick import FenwickTree, SegmentTree
 from .sde import euler_maruyama, milstein, gbm_paths
 from .grid_interp import bilinear_interp, nearest_interp
+from .robust_scale import qn_scale, sn_scale, biweight_midvariance
 from .levenberg import levenberg_marquardt
 from .ewma_cov import ewma_covariance_matrix, ewma_correlation_matrix
 from .vol_convert import black_to_normal_vol, normal_to_black_vol
@@ -1140,7 +1141,7 @@ try:  # pragma: no cover - trivial availability branch
 except ImportError:  # pragma: no cover
     HAS_NUMPY = False
 
-__version__ = "3.10.0"
+__version__ = "3.11.0"
 
 __all__ = [
     "OptionType",
@@ -2166,6 +2167,9 @@ __all__ = [
     "gbm_paths",
     "bilinear_interp",
     "nearest_interp",
+    "qn_scale",
+    "sn_scale",
+    "biweight_midvariance",
     "levenberg_marquardt",
     "ewma_covariance_matrix",
     "ewma_correlation_matrix",